% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyMatrix.R
\name{adjacencyMatrix}
\alias{adjacencyMatrix}
\title{adjacencyMatrix}
\usage{
adjacencyMatrix(
  graph_object,
  source_color = NULL,
  target_color = NULL,
  cellColor = NULL,
  columnBrush = TRUE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{graph_object}{microbiome composition graph build with build_mc_graph().}

\item{source_color}{color variable to set color on row labels (samples).}

\item{target_color}{color variable to set color on column labels (ASVs).}

\item{cellColor}{color variable to set color of cells.}

\item{columnBrush}{add barcode visualization as 2D brush to zoom in on ASVs/OTUs. Boolean.}

\item{width}{width of the graph. default is full available width.}

\item{haight}{height of the graph. default is full available height.}
}
\value{
interactive figure
}
\description{
adjacency matrix visualization of the microbiome composition graph
}
\examples{
# Default
adjacencyMatrix(graph_object = mcg)
# Color cells by relative abundance
adjacencyMatrix(graph_object = mcg, cellColor = "relative")
# Color sample labels by group
adjacencyMatrix(graph_object = mcg, cellColor = "relative", source_color = "group")
# Disable horizontal brushing
adjacencyMatrix(graph_object = mcg, cellColor = "relative", columnBrush = FALSE)

}
