% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R
\name{barcode}
\alias{barcode}
\title{barcode}
\usage{
barcode(graph_object, color = NULL, width = NULL, height = NULL)
}
\arguments{
\item{graph_object}{microbiome composition graph build with build_mc_graph().}

\item{color}{color variable to set color on cells.}

\item{width}{width of the graph. default is full available width.}

\item{haight}{hieght of the graph. default is full available height.}
}
\value{
interactive figure
}
\description{
barcode visualization of the microbiome composition graph that shows all observed ASVs/OTUs in the data.
}
\examples{
barcode(graph_object = mcg, color = "black")

}
