% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_graph.R
\name{build_mc_graph}
\alias{build_mc_graph}
\title{build_mcg}
\usage{
build_mc_graph(
  ps_object,
  beta_dist = "bray",
  seriation_method = NULL,
  seriation_dist = "bray"
)
}
\arguments{
\item{ps_object}{phyloseq object containing an 'otu_tab' and 'tax_table'.}

\item{beta_dist}{distance metric to be used to calculate pairwise distances between samples.}

\item{seriation_method}{seriation algorithm to calculate optimal order of rows and columns in adjacency matrix. See 'seriation' package for available options.}

\item{seriation_dist}{distance metric to be used to calculate pairwise distances between samples, and to be used as input for the seriation algorithm.}
}
\value{
A list containing two dataframes (nodes and links)
}
\description{
transforms an OTU or ASV data table stored as a phyloseq object into a microbiome composition graph (multivaraite bipartite graph).
}
\examples{
mcg <- build_mc_graph(ps_object = ps, beta_dist = "bray", seriation_method = NULL)

}
