% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceGraph.R
\name{forceGraph}
\alias{forceGraph}
\title{forceGraph}
\usage{
forceGraph(
  graph_object,
  source_color = "type",
  target_color = "type",
  groupingVariable = NULL,
  forceOnGroup = FALSE,
  nodeStroke = "black",
  nodeStrokeWidth = 1.5,
  nodeRadius_source = 10,
  nodeRadius_target = 4,
  linkStrokeWidth = 1.5,
  linkStroke = "gray",
  linkStrokeOpacity = 0.6,
  nodeOpacity = 1,
  nodeStrokeOpacity = 1,
  nodeTitle = NULL,
  highlightAbundance = TRUE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{graph_object}{microbiome composition graph build with build_mc_graph().}

\item{source_color}{color variable to set color on source nodes (samples).}

\item{target_color}{color variable to set color on target nodes (ASVs).}

\item{groupingVariable}{grouping variable between the samples.}

\item{forceOnGroup}{add repulsion force on groups. Boolean.}

\item{nodeStroke}{color of node stroke. defualt is "black". custum mapping can be done by providing an array in the following format c('variable', 'scaleType', 'domain', 'range', 'unknown').}

\item{nodeStrokeWidth}{width of the node stroke. default is 1.5.}

\item{nodeRadius_source}{radius of source nodes. default is 10. custum mapping can be done by providing an array in the following format c('variable', 'scaleType', 'domain', 'range', 'unknown').}

\item{nodeRadius_target}{radius of target nodes. default is 4. custum mapping can be done by providing an array in the following format c('variable', 'scaleType', 'domain', 'range', 'unknown').}

\item{linkStrokeWidth}{width of lines between connected nodes. default is 1.5. custum mapping can be done by providing an array in the following format c('variable', 'scaleType', 'domain', 'range', 'unknown').}

\item{linkStroke}{color of lines between nodes. default is "gray". custum mapping can be done by providing an array in the following format c('variable', 'scaleType', 'domain', 'range', 'unknown').}

\item{linkStrokeOpacity}{opacity of lines between nodes. default is 0.6. custum mapping can be done by providing an array in the following format c('variable', 'scaleType', 'domain', 'range', 'unknown').}

\item{nodeOpacity}{opacity of nodes. default is 1.}

\item{nodeStrokeOpacity}{opacity of nodes' stroke. default is 1.}

\item{nodeTitle}{array of variables you want to display in the tooltip. c('variable1','variable2')}

\item{highlightAbundance}{show taxa abundance within sample when hovering over a sample node and show taxon abundance in samples when hovering over a taxon node. Boolean.}

\item{width}{width of the graph. default is full available width.}

\item{height}{hieght of the graph. default is full available height.}
}
\value{
interactive figure
}
\description{
node-link visualization of the microbiome composition graph
}
\examples{
# Default
forceGraph(graph_object = mcg)
# Color source and target nodes
forceGraph(graph_object = mcg, source_color = "alphaShannon", target_color = "abundance")
# Set the node attributes to be printed while hovering on a node
forceGraph(graph_object = mcg, nodeTitle = c("id", "group"))
# Map relative abundance to edge color
forceGraph(graph_object = mcg, linkStroke = list("relative", "linear", c(0,0.1), c("grey", "blue")))
# Set grouping variable
forceGraph(graph_object = mcg, groupingVariable = "group")
# Disable highlighting of ASV abundance in samples
forceGraph(graph_object = mcg, highlightAbundance = FALSE)

}
